Jaime/jmx

+++ ci/common.rb
@@ -38,6 +38,83 @@ def install_requirements(req_file, pip_options = nil, output = nil, use_venv = n
   end
 end
 
+def travis_pr?
+  !ENV['TRAVIS'].nil? && ENV['TRAVIS_EVENT_TYPE'] == 'pull_request'
+end
+
+def wait_on_docker_logs(c_name, max_wait, *include_array)
+  count = 0
+  logs = `docker logs #{c_name} 2>&1`
+  puts "Waiting for #{c_name} to come up"
+
+  until count == max_wait || include_array.any? { |phrase| logs.include?(phrase) }
+    sleep(1)
+    logs = `docker logs #{c_name} 2>&1`
+    count += 1
+  end
+
+  if include_array.any? { |phrase| logs.include?(phrase) }
+    puts "#{c_name} is up!"
+  else
+    sh %(docker logs #{c_name} 2>&1)
+    raise
+  end
+end
+
+# Dict converting check.d name to Travis flavor names
+BAD_CITIZENS = {
+  'couch' => 'couchdb',
+  'disk' => 'system',
+  'dns_check' => 'system',
+  'network' => 'system',
+  'tcp_check' => 'system',
+  'http_check' => 'system',
+  'sysstat' => 'system',
+  'elastic' => 'elasticsearch',
+  'gearmand' => 'gearman',
+  'mcache' => 'memcache',
+  'php_fpm' => 'phpfpm',
+  'redisdb' => 'redis',
+  'ssh_check' => 'ssh',
+  'zk' => 'zookeeper'
+}.freeze
+
+def translate_to_travis(checks)
+  checks.map do |check_name|
+    check_name = BAD_CITIZENS[check_name] if BAD_CITIZENS.key? check_name
+    check_name
+  end
+end
+
+# rubocop:disable Metrics/AbcSize
+# rubocop:disable Metrics/MethodLength
+def can_skip?
+  return false, [] unless travis_pr?
+
+  modified_checks = []
+  puts "Comparing #{ENV['TRAVIS_PULL_REQUEST_SHA']} with #{ENV['TRAVIS_BRANCH']}"
+  git_output = `git diff --name-only #{ENV['TRAVIS_BRANCH']}...#{ENV['TRAVIS_PULL_REQUEST_SHA']}`
+  puts "Git diff: \n#{git_output}"
+  git_output.each_line do |filename|
+    filename.strip!
+    puts filename
+    if filename.start_with? 'checks.d'
+      check_name = File.basename(filename, '.py')
+    elsif filename.start_with?('tests/checks/integration', 'tests/checks/mock')
+      # 5 is test_
+      check_name = File.basename(filename, '.py').slice(5, 100)
+    elsif filename.start_with?('tests/checks/fixtures', 'conf.d')
+      next
+    else
+      return false, []
+    end
+    modified_checks << check_name unless modified_checks.include? check_name
+  end
+  [true, translate_to_travis(modified_checks)]
+end
+# rubocop:enable Metrics/AbcSize
+# rubocop:enable Metrics/MethodLength
+
 # helper class to wait for TCP/HTTP services to boot
 class Wait
   DEFAULT_TIMEOUT = 10
@@ -142,7 +219,9 @@ namespace :ci do
 
     task :before_cache do |t|
       section('BEFORE_CACHE')
-      sh %(find #{ENV['INTEGRATIONS_DIR']}/ -type f -name '*.log*' -delete)
+      unless Gem.win_platform?
+        sh %(find #{ENV['INTEGRATIONS_DIR']}/ -type f -name '*.log*' -delete)
+      end
       t.reenable
     end
 
@@ -163,8 +242,10 @@ namespace :ci do
 
       tests_directory = if flavors.include?('default') || flavors.include?('core_integration')
                           'tests/core'
-                        else
+                        elsif flavors.include?('checks_mock') || flavors.include?('windows')
                           'tests/checks'
+                        else
+                          flavors.map { |d| "tests/#{d}" }.compact.join(' ')
                         end
       # Rake on Windows doesn't support setting the var at the beginning of the
       # command
@@ -178,6 +259,41 @@ namespace :ci do
       sh %(#{path}nosetests -s -v -A "#{nose}" #{tests_directory})
       t.reenable
     end
-    task execute: [:before_install, :install, :before_script, :script]
+
+    task :execute, :flavor do |_t, attr|
+      flavor = attr[:flavor]
+      # flavor.scope.path is ci:cassandra
+      # flavor.scope.path[3..-1] is cassandra
+      travis_flavor = flavor.scope.path[3..-1]
+
+      can_skip, checks = can_skip?
+      can_skip &&= !%w[default core_integration checks_mock].include?(travis_flavor)
+
+      puts "Travis flavor: #{travis_flavor}"
+      puts "Detected modified checks: #{checks.join(' | ')}"
+
+      if can_skip && !checks.include?(travis_flavor)
+        puts "Skipping #{travis_flavor} tests, not affected by the change".yellow
+        next
+      end
+      exception = nil
+      begin
+        tasks = %w[before_install install before_script script]
+        tasks << 'before_cache' unless ENV['CI'].nil?
+        tasks.each do |t|
+          Rake::Task["#{flavor.scope.path}:#{t}"].invoke
+        end
+      rescue => e
+        exception = e
+        puts "Failed task: #{e.class} #{e.message}".red
+      end
+      if ENV['SKIP_CLEANUP']
+        puts 'Skipping cleanup, disposable environments are great'.yellow
+      else
+        puts 'Cleaning up'
+        Rake::Task["#{flavor.scope.path}:cleanup"].invoke
+      end
+      raise exception if exception
+    end
   end
 end
